Removed/conslidated Ansible overrides
