+++ provider/ansible.rb
@@ -18,11 +18,8 @@ require 'provider/ansible/documentation'
 require 'provider/ansible/example'
 require 'provider/ansible/manifest'
 require 'provider/ansible/module'
-require 'provider/ansible/property_override'
 require 'provider/ansible/request'
 require 'provider/ansible/resourceref'
-require 'provider/ansible/resource_override'
-require 'provider/ansible/property_override'
 require 'provider/ansible/facts_override'
 require 'provider/overrides/ansible/resource_override'
 require 'provider/overrides/ansible/property_override'
